# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LibRESTApi",
        "version": "1.0"
    },
    "paths": {
        "/api/": {},
        "/api/v1/book": {
            "get": {
                "tags": [
                    "Api"
                ],
                "summary": "Список",
                "operationId": "122f2c61509ebcc0f49cdc1b2c74e885",
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "5"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "The Picture of Dorian Gray"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "Oscar Wilde"
                                                    },
                                                    "year": {
                                                        "type": "string",
                                                        "example": "1890"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Api"
                ],
                "summary": "Создание",
                "operationId": "9bbcb64d84d7ab9b553fbc16faba8505",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "The Picture of Dorian Gray"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Oscar Wilde"
                                            },
                                            "year": {
                                                "type": "string",
                                                "example": "1890"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "5"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "The Picture of Dorian Gray"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Oscar Wilde"
                                                },
                                                "year": {
                                                    "type": "string",
                                                    "example": "1890"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/book/{book}": {
            "get": {
                "tags": [
                    "Api"
                ],
                "summary": "Получение записи по id",
                "operationId": "8e0ddddf5bea96241522354685bb2757",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "5"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "The Picture of Dorian Gray"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Oscar Wilde"
                                                },
                                                "year": {
                                                    "type": "string",
                                                    "example": "1890"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Api"
                ],
                "summary": "Удаление записи по id",
                "operationId": "f865c5c998701b113e59b3b952a830d6",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book removed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Api"
                ],
                "summary": "Обновление записи",
                "operationId": "47c7863c76402a341545e288ef283be6",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 5
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "The Picture of Dorian Gray"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Oscar Wilde"
                                            },
                                            "year": {
                                                "type": "string",
                                                "example": "1890"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "5"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "The Picture of Dorian Gray"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Oscar Wilde"
                                                },
                                                "year": {
                                                    "type": "string",
                                                    "example": "1890"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/book/{book}/reserve": {
            "post": {
                "tags": [
                    "Api"
                ],
                "summary": "Бронирование книги",
                "operationId": "e6a1198d6d7ce70bc8eeafdcf4e3fa35",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/book/{book}/unreserve": {
            "post": {
                "tags": [
                    "Api"
                ],
                "summary": "Возврат книги",
                "operationId": "3f957efe8f920128427ffc4bfd879d05",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Api",
            "description": "Api"
        }
    ]
}